// src/redux/tag-types.ts
export enum tagTypes {
  user = "user",
  admin = "admin",
  host = "host",
  soloRoom = "soloRoom",
  bookSoloRoom = "bookSoloRoom",
  package = "package",
  bookPackage = "bookPackage",
  packageReviewRating = "packageReviewRating",
  soloRoomReviewRating = "soloRoomReviewRating",
}

export const tagTypesList = [
  tagTypes.user,
  tagTypes.admin,
  tagTypes.host,
  tagTypes.soloRoom,
  tagTypes.bookSoloRoom,
  tagTypes.package,
  tagTypes.bookPackage,
  tagTypes.packageReviewRating,
  tagTypes.soloRoomReviewRating,
];